requirements.txt
attrs==19.3.0
dataclasses==0.7
importlib-metadata==1.5.0
jellyfish==0.7.2
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
prettytable==0.7.2
prompt-toolkit==3.0.4
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.4.1
python-nubia==0.2b2
six==1.14.0
termcolor==1.1.0
wcwidth==0.1.8
zipp==3.1.0
